pyproject.toml
[tool.poetry]
name = "birthday-wisher"
version = "0.1.0"
description = ""
authors = ["Farhadur Fahim"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
